Docker build update.
Both docker files are updated.

Also use the `Env` module to set environment variables in
`lib/RenderApp.pm`.  There is more work needed here.  The environment
variables that are set from config variables should not be.  It looks
like this was done as a quick hack to make these values available
everywhere in the code.  Instead the values should be obtained from the
config when needed, and passed in as parameters when used outside of the
Mojolicious app areas.

Also, don't set the `MOJO_LOG_LEVEL` environment variable by default.
Let the user set that if desired. Using a hard coded log level is not a
good idea.  This could also be changed to be set from a config variable
for convenience.